requirements.txt
langchain==0.3.9
langchain_community==0.3.9
langchain_openai==0.2.11
openai==1.56.2
python-dotenv==1.0.1
streamlit==1.40.2
